+++ CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org).
 
+## [5.0.1] - 2024-11-28
+
+## Changed
+
+- Gemmi fractionalization_matrix to frac.mat and orthogonalization_matrix to orth.mat.
+
+## [5.0.0] - 2024-11-08
+
+### Added
+
+- EM --half-maps, --single-map and --build-map arguments.
+
+### Removed
+
+- EM --map and --blur arguments.
+
+## [4.0.2] - 2024-10-24
+
+### Fixed
+
+- Crash when no residues are built.
+
 ## [4.0.1] - 2024-05-14
 
 ### Fixed